requirements.txt
boto3==1.28.40
decouple
numpy==1.24.4
pandas==2.1.0
pyarrow==13.0.0
torch==2.0.1
soundfile==0.12.1
pydub==0.25.1
openai-whisper
python-dotenv


